add relationship between models

+++ src/controllers/comments-controller.go
@@ -8,6 +8,8 @@ import (
 	"go-api/src/repositories"
 )
 
+const commentNotFoundMsg = "Comment not found"
+
 func GetAllComments(c *gin.Context) {
 	db := c.MustGet("DB").(*gorm.DB)
 
@@ -33,10 +35,10 @@ func GetCommentById(c *gin.Context) {
 
 	db := c.MustGet("DB").(*gorm.DB)
 	commentRepository := repositories.NewCommentRepository(db)
-	comment, err := commentRepository.GetCommentByID(uint(commentID))
-
 	if err != nil {
-		c.JSON(404, gin.H{"error": "Comment not found"})
+		c.JSON(404, gin.H{"error": commentNotFoundMsg})
+		return
+	}
 		return
 	}
 
@@ -53,6 +55,24 @@ func AddComment(c *gin.Context) {
 
 	db := c.MustGet("DB").(*gorm.DB)
 	commentRepository := repositories.NewCommentRepository(db)
+
+	user := c.MustGet("User").(models.User)
+	comment.UserID = user.ID
+	comment.User = user
+
+	languageID, err := strconv.ParseUint(c.Query("language_id"), 10, 32)
+	if err != nil || languageID <= 0 {
+		c.JSON(400, gin.H{"error": "Invalid language ID"})
+		return
+	}
+	comment.LanguageID = uint(languageID)
+	languageRepository := repositories.NewLanguageRepository(db)
+	language, err := languageRepository.GetLanguageByID(comment.LanguageID)
+	if err != nil {
+		c.JSON(404, gin.H{"error": "Language not found"})
+		return
+	}
+	comment.Language = language
 
 	if err := commentRepository.CreateComment(&comment); err != nil {
 		c.JSON(500, gin.H{"error": "Failed to create comment"})
@@ -74,8 +94,19 @@ func DeleteComment(c *gin.Context) {
 	db := c.MustGet("DB").(*gorm.DB)
 	commentRepository := repositories.NewCommentRepository(db)
 
+	user := c.MustGet("User").(models.User)
+	comment, err := commentRepository.GetCommentByID(uint(commentID))
+	if err != nil {
+		c.JSON(404, gin.H{"error": commentNotFoundMsg})
+		return
+	}
+	if comment.UserID != user.ID {
+		c.JSON(403, gin.H{"error": "You are not authorized to delete this comment"})
+		return
 	if err := commentRepository.DeleteComment(uint(commentID)); err != nil {
-		c.JSON(404, gin.H{"error": "Comment not found"})
+		c.JSON(404, gin.H{"error": commentNotFoundMsg})
+		return
+	}
 		return
 	}
